+++ CHANGELOG.md
@@ -13,6 +13,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 * Feature - User Management - App Access Table Column Heading Change
 * Feature - App Release Process
    * Update of Step 4 to only handle the role upload
+   * Beta Test- Added preview content
 * BugFix - Add user role for app not working fix
 * Updates
    * Style fix in Company roles page